requirements.txt
sqlalchemy==1.4.36
pyTelegramBotAPI==4.10.0
python-dotenv